APPBackend/Django/requirements.txt
Django==1.11.5
django-background-tasks==1.1.10
django-compat==1.0.14
djangorestframework==3.6.3
jsonfield==2.0.2
pytz==2017.2
six==1.10.0
virtualenvwrapper==4.7.2
psycopg2==2.7.0
celery
redis
django-tarview
docker
gunicorn
djangorestframework-jwt
